much faster VIS merging
Also found a potential bug from the first speedup. Something is wrong with the 64bit shifting but I don't want to find out what, so it's back to byte-size shifts. Still much faster due to the memcpy stuff.

+++ src/BspMerger.h
@@ -65,7 +65,7 @@ private:
 
 	void print_merge_progress();
 
-	bool shiftVis(uint64* vis, int len, int offsetLeaf, int shift);
+	bool shiftVis(byte* vis, int len, int offsetLeaf, int shift);
 
 	// decompress this map's vis data into arrays of bits where each bit indicates if a leaf is visible or not
 	// iterationLeaves = number of leaves to decompress vis for
